Move `PFHRGBSignature250` to feature types header
