+++ features/include/pcl/features/pfhrgb.h
@@ -39,6 +39,7 @@
 
 #pragma once
 
+#include <pcl/feature_types.h>
 #include <pcl/features/feature.h>
 
 namespace pcl
